AI chat/AI Curhat/api/requirements.txt
Flask
google-cloud-language
gunicorn
python-dotenv
